Refactoring of the propagator
This commit introduce a big refactor of the code base. The new
propagator works on the clauses and detect which one are T-reachable and
F-reachable.
Hence the Graph structure have been modified such that the nodes of the
graph are clauses.
In addition a new caching system have been implemented that correctly
handle hash collision, which wasn't the case previously but the
probablity of a collision was very low.
Tests need to be added.