Refactoring the code
The code has been massively refactored for a few reason. The main one is
that I did not like how the code was organized and the addition of
clause learning was a pain.

Basically the clauses are now uniformed as a set of literals. It has an
optional head for Horn clauses (input of the problem).
A lot of things that were in grap.rs are now moved into their own files,
which increase the lisiblity of the code and makes it easier to
unit-test.

The search has been unified as the exact search can be viewed as an
approximate search with epsilon = 0.
The propagator is not const-generic anymore.

These changes might induce a small overhead, but it should not be that
visible.